+++ src/frequenz/repo/config/pytest/__init__.py
@@ -1,0 +1,12 @@
+# License: MIT
+# Copyright © 2023 Frequenz Energy-as-a-Service GmbH
+
+"""Pytest utilities.
+
+This package contains utilities for testing with [`pytest`](https://pypi.org/project/pytest/).
+
+The following modules are available:
+
+- [`examples`][frequenz.repo.config.pytest.examples]: Utilities to enable linting of
+  code examples in docstrings.
+"""